+++ src/Disks/S3/registerDiskS3.cpp
@@ -141,7 +141,7 @@ void registerDiskS3(DiskFactory & factory)
             metadata_path,
             context.getSettingsRef().s3_min_upload_part_size,
             config.getUInt64(config_prefix + ".min_multi_part_upload_size", 10 * 1024 * 1024),
-            config.getUInt64(config_prefix + ".read_seek_threshold", 1024 * 1024));
+            config.getUInt64(config_prefix + ".min_bytes_for_seek", 1024 * 1024));
 
         /// This code is used only to check access to the corresponding disk.
         checkWriteAccess(*s3disk);
