+++ src/IO/ya.make
@@ -39,6 +39,7 @@ SRCS(
     readFloatText.cpp
     ReadHelpers.cpp
     ReadWriteBufferFromHTTP.cpp
+    SeekAvoidingReadBuffer.cpp
     UseSSL.cpp
     WriteBufferAIO.cpp
     WriteBufferFromFile.cpp
